requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
psycopg2-binary
celery
redis
python-dotenv
google-genai
python-multipart


